Call auth callback after Web View is closed
Before the callback would just be sent to the extractCode method. Now
this method gets a new callback, which checks if we are using a web
view. If we are this gets dismissed, and the original callback is passed
to the dismissViewControllerAnimated callback. That way the callback
will be called last, after the Web View is closed.